+++ src/styles/components/_articleImageInput.scss
@@ -12,4 +12,52 @@
         height: 5rem;
         font-size: 2rem;
     }
+
+    &__upload-image-wrap {
+        display: flex;
+        flex-direction: column;
+        margin-bottom: 3rem;
+    }
+
+    &__upload-image-list {
+        display: flex;
+        align-items: center;
+    }
+
+    &__upload-image-item {
+        position: relative;
+        margin-right: 1rem;
+        width: 5rem;
+        height: 5rem;
+
+        & svg {
+            position: absolute;
+            top: -5px;
+            right: -5px;
+            background-color: white;
+            border: none;
+            border-radius: 50%;
+            font-size: 1rem;
+            color: black;
+            z-index: 1;
+            cursor: pointer;
+        }
+    }
+
+    &__upload-image {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+        border-radius: 0.2rem;
+    }
+
+    &__upload-image-length {
+        align-self: end;
+        font-weight: 700;
+        color: $primary-dark;
+
+        &--limit {
+            color: $secondary-dark;
+        }
+    }
 }
