+++ src/styles/pages/_createCommunity.scss
@@ -26,7 +26,7 @@
         }
     }
 
-    &__title-length, &__upload-image-length {
+    &__title-length {
         align-self: end;
         font-weight: 700;
         color: $primary-dark;
@@ -34,44 +34,6 @@
         &--limit {
             color: $secondary-dark;
         }
-    }
-
-    &__upload-image-wrap {
-        display: flex;
-        flex-direction: column;
-        margin-bottom: 3rem;
-    }
-
-    &__upload-image-list {
-        display: flex;
-        align-items: center;
-    }
-
-    &__upload-image-item {
-        position: relative;
-        margin-right: 1rem;
-        width: 5rem;
-        height: 5rem;
-
-        & svg {
-            position: absolute;
-            top: -5px;
-            right: -5px;
-            background-color: white;
-            border: none;
-            border-radius: 50%;
-            font-size: 1rem;
-            color: black;
-            z-index: 1;
-            cursor: pointer;
-        }
-    }
-
-    &__upload-image {
-        width: 100%;
-        height: 100%;
-        object-fit: cover;
-        border-radius: 0.2rem;
     }
 
     &__button-wrap {
